board 수정, 삭제 기능 추가 - 방어코드 추가 - 실제로 삭제하지않고 deleted 0, 1로 조절 - deleted 1인 게시물은 리스트에 표시하지않게 수정

+++ src/main/java/com/gameduos/springboot/web/controller/BoardRestController.java
@@ -43,15 +43,26 @@ public class BoardRestController {
         return boardService.save(requestDto);
     }
 
-
     @PutMapping("/{idx}")
-    public ResponseEntity<?> putBoard(@PathVariable("idx")Long idx, @RequestBody BoardUpdateRequestDto requestDto) {
+    public ResponseEntity<?> putBoard(@PathVariable("idx")Long idx, @RequestBody BoardUpdateRequestDto requestDto,
+                                      @SocialUser User user) {
+        User boardAuthor = boardService.findBoardUserById(idx);
+
+        if(!user.getId().equals(boardAuthor.getId())){
+            throw new IllegalStateException("자신의 게시글만 수정할 수 있습니다.");
+        }
+
         return boardService.update(idx, requestDto);
     }
 
-
     @DeleteMapping("/{idx}")
-    public ResponseEntity<?> deleteBoard(@PathVariable("idx")Long idx) {
+    public ResponseEntity<?> deleteBoard(@PathVariable("idx")Long idx, @SocialUser User user) {
+        User boardAuthor = boardService.findBoardUserById(idx);
+
+        if(!user.getId().equals(boardAuthor.getId())){
+            throw new IllegalStateException("자신의 게시글만 삭제할 수 있습니다.");
+        }
+
         return boardService.delete(idx);
     }
 }
